refator: separate delete post situation to a different handlers

+++ bot/handlers/admin.py
@@ -93,38 +93,64 @@ async def post_queue_navigation(callback: CallbackQuery, callback_data: Navigati
     
     await callback.answer()
 
-@router.callback_query(AdminPanel.post_queue_page, DeletePostCB.filter())
-async def delete_post_btn_clicked(callback: CallbackQuery, callback_data: DeletePostCB, state: FSMContext):
+@router.callback_query(
+    AdminPanel.post_queue_page, 
+    DeletePostCB.filter(F.source == "list")
+)
+async def delete_from_list(callback: CallbackQuery, callback_data: DeletePostCB, state: FSMContext):
     post_id = callback_data.id
     current_page = callback_data.page
 
     await service.delete_post(post_id)
 
+    state_data = await state.get_data()
+    opened_post_id = state_data.get("opened_post_id")
+    view_msg_id = state_data.get("opened_post_msg_id")
+
+    if view_msg_id and opened_post_id == post_id:
+        with suppress(TelegramBadRequest):
+            await callback.message.bot.delete_message(
+                chat_id=callback.message.chat.id,
+                message_id=view_msg_id
+            )
+        await state.update_data(opened_post_msg_id=None, opened_post_id=None)
+
+    message_text, reply_markup = await window.get_post_queue_window(current_page)
+    
+    with suppress(TelegramBadRequest):
+        await callback.message.edit_text(text=message_text, reply_markup=reply_markup)
+
     await callback.answer("Пост удален")
 
-    state_data = await state.get_data()
-    view_msg_id = state_data.get("opened_post_msg_id")
-    opened_post_id = state_data.get("opened_post_id")
-    list_msg_id = state_data.get("list_msg_id")
-    
-    if view_msg_id:
-        is_deleting_opened_post_from_list = opened_post_id == post_id
-        is_deleting_from_opened_post_btn = callback.message.message_id == view_msg_id
-        if is_deleting_opened_post_from_list or is_deleting_from_opened_post_btn:
-            with suppress(TelegramBadRequest):
-                await callback.message.bot.delete_message(
-                    chat_id=callback.message.chat.id,
-                    message_id=view_msg_id
-                )
-    
-    message_text, reply_markup = await window.get_post_queue_window(current_page)
+@router.callback_query(
+    AdminPanel.post_queue_page, 
+    DeletePostCB.filter(F.source == "view")
+)
+async def delete_from_view(callback: CallbackQuery, callback_data: DeletePostCB, state: FSMContext):
+    post_id = callback_data.id
+    current_page = callback_data.page
 
-    with suppress(TelegramBadRequest):
-        await callback.message.bot.edit_message_text(
-            chat_id=callback.message.chat.id,
-            message_id=list_msg_id,
-            text=message_text,
-            reply_markup=reply_markup)    
+    await service.delete_post(post_id)
+    
+    await state.update_data(opened_post_msg_id=None, opened_post_id=None)
+
+    state_data = await state.get_data()
+    list_msg_id = state_data.get("list_msg_id")
+
+    if list_msg_id:
+        message_text, reply_markup = await window.get_post_queue_window(current_page)
+        
+        with suppress(TelegramBadRequest):
+            await callback.message.bot.edit_message_text(
+                chat_id=callback.message.chat.id,
+                message_id=list_msg_id,
+                text=message_text,
+                reply_markup=reply_markup
+            )
+
+    await callback.answer("Пост удален", show_alert=True)
+
+    await callback.message.delete()
 
 @router.callback_query(
         AdminPanel.post_queue_page,
